# bounds check elimation in main loop of encryption