--- a/sm4/sm4.go
+++ b/sm4/sm4.go
@@ -184,7 +184,6 @@ func permuteFinalBlock(b []byte, block []uint32) {
 
 //修改后的加密核心函数
 func cryptBlock(subkeys []uint32, b []uint32, r []byte, dst, src []byte, decrypt bool) {
-	var x uint32
 	permuteInitialBlock(b, src)
 
 	// bounds check elimination in major encryption loop
@@ -193,7 +192,7 @@ func cryptBlock(subkeys []uint32, b []uint32, r []byte, dst, src []byte, decrypt
 	if decrypt {
 		for i := 0; i < 8; i++ {
 			s := subkeys[31-4*i-3 : 31-4*i-3+4]
-			x = b[1] ^ b[2] ^ b[3] ^ s[3]
+			x := b[1] ^ b[2] ^ b[3] ^ s[3]
 			b[0] = b[0] ^ sbox0[x&0xff] ^ sbox1[(x>>8)&0xff] ^ sbox2[(x>>16)&0xff] ^ sbox3[(x>>24)&0xff]
 			x = b[0] ^ b[2] ^ b[3] ^ s[2]
 			b[1] = b[1] ^ sbox0[x&0xff] ^ sbox1[(x>>8)&0xff] ^ sbox2[(x>>16)&0xff] ^ sbox3[(x>>24)&0xff]
@@ -205,7 +204,7 @@ func cryptBlock(subkeys []uint32, b []uint32, r []byte, dst, src []byte, decrypt
 	} else {
 		for i := 0; i < 8; i++ {
 			s := subkeys[4*i : 4*i+4]
-			x = b[1] ^ b[2] ^ b[3] ^ s[0]
+			x := b[1] ^ b[2] ^ b[3] ^ s[0]
 			b[0] = b[0] ^ sbox0[x&0xff] ^ sbox1[(x>>8)&0xff] ^ sbox2[(x>>16)&0xff] ^ sbox3[(x>>24)&0xff]
 			x = b[0] ^ b[2] ^ b[3] ^ s[1]
 			b[1] = b[1] ^ sbox0[x&0xff] ^ sbox1[(x>>8)&0xff] ^ sbox2[(x>>16)&0xff] ^ sbox3[(x>>24)&0xff]